+++ CHANGELOG.md
@@ -47,3 +47,8 @@ Now plugin always must receive `params` object
 * Fix of incorrect defaults for `LikesPluginLikesTable`
 * Return nullable (not required) parameter for plugin
 * Update version of base bot
+
+### 0.8.6
+
+* Fix attach like
+* Add refresh message marks